Single components properties file created
